ScenarioGenerator/requirements.txt
gunicorn==20.0.4
psycopg2-binary==2.8.5
Django==3.0.7
djangorestframework==3.12.4
Markdown==3.3.4
numpy==1.19.5
paramiko==2.7.2
protobuf==3.14.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pytz==2021.1
scenariogeneration==0.3.4
scipy==1.6.3
six==1.15.0
sqlparse==0.4.1
requests